% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_split.R
\name{exercise_split}
\alias{exercise_split}
\title{exercise_split}
\usage{
exercise_split(
  inputdirectory,
  outputdirectory,
  exercisefile,
  hourspostexercise
)
}
\arguments{
\item{inputdirectory}{path to folder containing cleaned CGM files. See README on structure for these files follwoing the output of CleanCGM}

\item{outputdirectory}{path to folder where new folders will be created}

\item{exercisefile}{path to file containing list of exercise timestamps for every person in study. See README on structure for this file.}

\item{hourspostexercise}{numeric specified for the time of interest post exercise, defaults to 6}
}
\value{
Returns folders containing individual files for the corresponding split CGM data per individual.
These files in each folder can then be run through the analyseCGM() function with the exercise parameter = TRUE to get CGM metrics within those specific periods and used for plotting CGM traces.
}
\description{
This function splits up clean CGM data into timeperiods specfic to post exercise analysis: default 6 hours (changeable) post exercise, 24 hours post exercise, overnight (0000-0600), postexercise-0000, next day post exercise (0600-2400)
This function accepts clean CGM data in the format of the output of cleanCGM(). See README for details)
}
\seealso{
analyseCGM and cleanCGM
}
\author{
Alice Carr
}
