% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_split.R
\name{intervention_split}
\alias{intervention_split}
\title{intervention_split}
\usage{
intervention_split(
  inputdirectory,
  aggregated = F,
  outputdirectory,
  interventionfile
)
}
\arguments{
\item{inputdirectory}{path to folder containing cleaned CGM files. See README on structure for these files follwoing the output of cleanCGM}

\item{aggregated}{TRUE/FALSE dictates if data is pre-aggregated or in individual separate files (usual for clinical trials)}

\item{outputdirectory}{path to folder where new folders will be created}

\item{interventionfile}{path to file containing list of intervention timestamps for every person in study. See README on structure for this file. Ids must match ids in CGM}
}
\value{
Returns folders containing individual (or pre aggregated depending on input structure) files for the corresponding split CGM data per individual.
}
\description{
This function splits up clean CGM data into specfic periods before and after intervention(s) based on intervention description file, to enable
This function accepts clean CGM data in the format of the output of cleanCGM(). (See README for details)
}
\seealso{
analyseCGM and cleanCGM
}
\author{
Alice Carr
}
