% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCGM.R
\name{analyseCGM}
\alias{analyseCGM}
\title{analyseCGM}
\usage{
analyseCGM(
  exercise = F,
  hourspostexercise = NULL,
  aggregated = F,
  analysesensorlifetime = T,
  libre = T,
  inputdirectory,
  outputdirectory,
  outputname = "CGMupload",
  belowexcursionlength = 15,
  aboveexcursionlength = 15,
  magedef = "1sd",
  congan = 1,
  format = "rows",
  printname = T
)
}
\arguments{
\item{exercise}{Default is FALSE. In house development of time windowed exercise files.
Format of these files is slightly different to the output files from \code{\link[=cleanCGM]{cleanCGM()}} function}

\item{hourspostexercise}{only use if exercise=T. Accepts a numeric for the specific restricted timeperiod post exercise (usually 6 hours)}

\item{aggregated}{TRUE/FALSE. Default is FALSE. This aims to handle pre aggregated data where more than one individual/visit etc. is in the file, most likely from external clinical study databases.}

\item{analysesensorlifetime}{TRUE/FALSE. Default is TRUE. Used only with aggregated= TRUE. If you are expecting to analyse continuous data that is > sensor lifetime then specify to FALSE. This is to ensure multiple sensors from one person are not aggregated by accident.}

\item{libre}{For calculation of the correct interval (libre 15 min or 900s, CGM 5 min or 300s).
Default is FALSE. Currently libre files have "dummy" coded 5 minute data with carry forward method.
ie. Every row there is an addition of 2 rows that are same as the original row.}

\item{inputdirectory}{path to folder containing files created by \code{\link[=cleanCGM]{cleanCGM()}} function}

\item{outputdirectory}{path to folder where output csv file will be uploaded}

\item{outputname}{name of output file. Default "CGMupload".}

\item{belowexcursionlength}{numeric for the time (in minutes) defined as an hypoglycemic exercusion
default is 15 minutes (https://care.diabetesjournals.org/content/40/12/1631)}

\item{aboveexcursionlength}{numeric for the time (in minutes) defined as an hyperglycemic exercusion
default is 15 minutes (https://care.diabetesjournals.org/content/40/12/1631)}

\item{magedef}{Defining the threshold used in MAGE calculation.  MAGE is an arithmetic average of either the upward or downward
of all glycemic excursions exceeding the threshold (standard deviation of blood glucose obtained from all blood glucose
concentrations within 24-hour period). Default is 1 standarddevation ("1sd"), options are 1.5 SD ("1.5sd") , 2 SD ("2sd")
or other can be specifed as a numeric}

\item{congan}{Specificing the n number of hours in CONGA(n). Default is the numeric 1. CONGA(n) represents the SD
of all valid differences between a current observation and an observation (n) hours earlier}

\item{format}{changes format to CGM variables as x or y in table. Default is "rows" making each ID a row}

\item{printname}{Prints name of the file being processed. Default is TRUE.}
}
\value{
A .csv file called by default CGMupload.csv containing CGM variables per each individual (file) on each row
}
\description{
This function function written to create consensus glycemic metrics based off definitions outlined International Consensus on Use of Continuous Glucose Monitoring. Ths function takes files of a certain structure as outlined in the README.
}
\seealso{
cleanCGM and exercise_split
}
\author{
Alice Carr
}
